making everything accesable from main
